`@remotion/renderer`: Cleanup headless browsers inbetween renders
