+++ packages/renderer/src/browser/Browser.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import {removeHeadlessBrowser} from '../browser-instances';
 import type {LogLevel} from '../log-level';
 import {assert} from './assert';
 import type {Page} from './BrowserPage';
@@ -278,6 +279,7 @@ export class HeadlessBrowser extends EventEmitter {
 		this.emit(
 			silent ? BrowserEmittedEvents.ClosedSilent : BrowserEmittedEvents.Closed,
 		);
+		removeHeadlessBrowser(this);
 	}
 
 	disconnect(): void {
