+++ packages/renderer/src/index.ts
@@ -43,7 +43,7 @@ import {DEFAULT_JPEG_QUALITY, validateJpegQuality} from './jpeg-quality';
 import {isEqualOrBelowLogLevel, isValidLogLevel, logLevels} from './log-level';
 import {INDENT_TOKEN, Log} from './logger';
 import {mimeContentType, mimeLookup} from './mime-types';
-import {internalOpenBrowser, killAllBrowsers} from './open-browser';
+import {internalOpenBrowser} from './open-browser';
 import {
 	DEFAULT_OPENGL_RENDERER,
 	validOpenGlRenderers,
@@ -132,6 +132,7 @@ export {validateOutputFilename} from './validate-output-filename';
 export type {AudioCodec};
 
 import {makeDownloadMap} from './assets/download-map';
+import {killAllBrowsers} from './browser-instances';
 import {codecSupportsMedia} from './codec-supports-media';
 import {makeFileExecutableIfItIsNot} from './compositor/make-file-executable';
 import {internalEnsureBrowser} from './ensure-browser';
